Expanded item table to take feats & effects
feats, effects and items are all merged into one table. A column named
"Type" defines what kind of item is in there. This seems to have
simplified much of the program code as we dont need an extra copy of
code for feats. Items are now added by type for buffs to the character.